refactor(scripts): separate introduce barrel api within /jest
