+++ apps/react-18-tests-v8/jest.config.js
@@ -1,29 +1,25 @@
 // @ts-check
-
-const { resolveMergeStylesSerializer } = require('@fluentui/scripts/jest/jest-resources');
-const { getResolveAlias } = require('@fluentui/scripts/webpack');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 /**
  * @type {import('@jest/types').Config.InitialOptions}
  */
-module.exports = {
+const config = createConfig({
   displayName: 'react-18-tests-v8',
-  preset: '../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsConfig: '<rootDir>/tsconfig.spec.json',
-      diagnostics: { warnOnly: true, exclude: ['packages/**'] },
-    },
-  },
-  transform: {
-    '^.+\\.tsx?$': 'ts-jest',
-  },
-
   coverageDirectory: './coverage',
   setupFilesAfterEnv: ['./config/tests.js'],
-  moduleNameMapper: {
-    '\\.(scss)$': '@fluentui/scripts/jest/jest-style-mock.js',
-    ...getResolveAlias(),
-  },
-  snapshotSerializers: [resolveMergeStylesSerializer()],
-};
+  snapshotSerializers: ['@fluentui/jest-serializer-merge-styles'],
+});
+
+if (config.globals) {
+  // override ts-jest config, otherwise it gets merged
+  config.globals['ts-jest'] = {
+    tsConfig: '<rootDir>/tsconfig.spec.json',
+    diagnostics: { warnOnly: true /* , exclude: ['packages/**']  */ },
+  };
+}
+
+// use default jest config to properly resolve react-18
+delete config.moduleDirectories;
+
+module.exports = config;
