+++ packages/date-time-utilities/jest.config.js
@@ -1,4 +1,4 @@
 // @ts-check
-const { createConfig } = require('@fluentui/scripts/jest/jest-resources');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 module.exports = createConfig({});
