+++ packages/jest-serializer-merge-styles/jest.config.js
@@ -1,5 +1,5 @@
 const path = require('path');
-const { createConfig } = require('@fluentui/scripts/jest/jest-resources');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 const config = createConfig({
   snapshotSerializers: [path.join(__dirname, 'src', 'index.ts')],
