+++ packages/merge-styles/jest.config.js
@@ -1,3 +1,3 @@
-let { createConfig } = require('@fluentui/scripts/jest/jest-resources');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 module.exports = createConfig();
