+++ packages/react-migration-v8-v9/jest.config.js
@@ -1,9 +1,8 @@
-let { createConfig, resolveMergeStylesSerializer } = require('@fluentui/scripts/jest/jest-resources');
-let path = require('path');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 const config = createConfig({
-  setupFiles: [path.resolve(path.join(__dirname, 'config', 'tests.js'))],
-  snapshotSerializers: [resolveMergeStylesSerializer(), 'enzyme-to-json/serializer'],
+  setupFiles: ['./config/tests.js'],
+  snapshotSerializers: ['@fluentui/jest-serializer-merge-styles', 'enzyme-to-json/serializer'],
 });
 
 module.exports = config;
