+++ packages/react-monaco-editor/jest.config.js
@@ -1,4 +1,4 @@
-const { createConfig } = require('@fluentui/scripts/jest/jest-resources');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 const config = createConfig({
   moduleNameMapper: {
