+++ packages/react/jest.config.js
@@ -1,9 +1,8 @@
-const { createConfig, resolveMergeStylesSerializer } = require('@fluentui/scripts/jest/jest-resources');
-const path = require('path');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 
 const config = createConfig({
-  setupFiles: [path.resolve(path.join(__dirname, 'config', 'tests.js'))],
-  snapshotSerializers: [resolveMergeStylesSerializer()],
+  setupFiles: ['./config/tests.js'],
+  snapshotSerializers: ['@fluentui/jest-serializer-merge-styles'],
 });
 
 module.exports = config;
