+++ packages/test-utilities/jest.config.js
@@ -1,2 +1,2 @@
-let { createConfig } = require('@fluentui/scripts/jest/jest-resources');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 module.exports = createConfig({});
