+++ tools/generators/migrate-converged-pkg/__fixtures__/old-jest-config.js__tmpl__
@@ -1,8 +1,8 @@
-const { createConfig } = require('@fluentui/scripts/jest/jest-resources');
+const { createV8Config: createConfig } = require('@fluentui/scripts/jest');
 const path = require('path');
 
 const config = createConfig({
-  setupFiles: [path.resolve(path.join(__dirname, 'config', 'tests.js'))],
+   setupFiles: ['./config/tests.js'],
   snapshotSerializers: ['@griffel/jest-serializer'],
 });
